+++ packages/design-system-web/.build/stencil/stencil.basic.config.ts
@@ -1,6 +1,6 @@
 import { Config } from '@stencil/core';
 import { angularOutputTarget } from '@stencil/angular-output-target';
-import { reactOutputTarget as react } from '@geovistory/react-output-target';
+import { reactOutputTarget } from '@stencil/react-output-target';
 import { sass } from '@stencil/sass';
 
 export const stencilBasicConfig: Config = {
@@ -28,7 +28,7 @@ export const stencilBasicConfig: Config = {
       directivesProxyFile: '../design-system-angular/projects/component-library/src/lib/stencil-generated/components.ts',
       directivesArrayFile: '../design-system-angular/projects/component-library/src/lib/stencil-generated/index.ts',
     }),
-    react({
+    reactOutputTarget({
       componentCorePackage: '@geovistory/design-system-web',
       proxiesFile: '../design-system-react/src/components/stencil-generated/index.ts',
       includeDefineCustomElements: true,